change from postgresql to mysql

+++ internal/domain/viewmodel/bet.go
@@ -1,7 +1,7 @@
 package viewmodel
 
 type Bet struct {
-	ID        int32     `json:"id,omitempty"`
+	ID        int64    `json:"id,omitempty"`
 	Gambler   *Gambler `json:"gambler,omitempty"`
 	BetType   string  `json:"bet_type,omitempty"`
 	BetPrice  float64 `json:"bet_price,omitempty"`
