+++ internal/domain/viewmodel/gambler.go
@@ -3,7 +3,7 @@ package viewmodel
 import "time"
 
 type Gambler struct {
-	ID           int32     `json:"id,omitempty"`
+	ID           int64     `json:"id,omitempty"`
 	Name         string    `json:"name,omitempty"`
 	Document     string    `json:"document,omitempty"`
 	DocumentType string    `json:"document_type,omitempty"`
